[UPDATE] update CRUD students, modify method delete parent, modify using StringUtils

+++ src/main/java/yoot/week1/service/impl/ParentServiceImpl.java
@@ -2,9 +2,11 @@ package yoot.week1.service.impl;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
+import org.springframework.util.StringUtils;
 import yoot.week1.domain.entity.Parent;
 import yoot.week1.dto.request.UpdateParentRequest;
 import yoot.week1.repository.ParentReposity;
+import yoot.week1.repository.StudentRepository;
 import yoot.week1.service.ParentService;
 
 import java.util.List;
@@ -14,6 +16,7 @@ import java.util.Optional;
 @RequiredArgsConstructor
 public class ParentServiceImpl implements ParentService {
     private final ParentReposity parentReposity;
+    private final StudentRepository studentRepository;
 
     public List<Parent> findAll(){
         return parentReposity.findAll();
@@ -24,10 +27,14 @@ public class ParentServiceImpl implements ParentService {
     }
 
     public Parent save(Parent parent){
+
         return parentReposity.save(parent);
     }
 
     public void delete(long id){
+        if(studentRepository.existsByParent_Id(id)){
+            throw new RuntimeException("Parent has student in yoot");
+        }
         parentReposity.deleteById(id);
     }
 
@@ -37,14 +44,14 @@ public class ParentServiceImpl implements ParentService {
 
     public Parent update(Long id, UpdateParentRequest request){
         Parent parent = findById(id).orElseThrow(() -> new RuntimeException("Parent not found"));
-        if (request.fullName()!=null){
+        if (StringUtils.hasText(request.fullName()) ){
             parent.setFullName(request.fullName());
         }
-        if (request.email()!=null){
+        if (StringUtils.hasText(request.email())){
             parent.setEmail(request.email());
         }
 
-        if(request.phone()!=null){
+        if(StringUtils.hasText(request.phone())){
             parent.setPhone(request.phone());
         }
